refactor(planning): extract 10 planner tools into a registry-based factory
Phase 7 of the cqrs refactor (#227), Task 2.

Each of the 10 root-planner custom tools moves from an inline
`createXxxTool()` method on `PlanningService` into its own file
under `src/modules/planning/tools/`:

  - graph-query.tool.ts
  - propose-mutations.tool.ts
  - graph-mutate.tool.ts
  - memory-read.tool.ts
  - memory-write.tool.ts
  - delegate-subagent.tool.ts
  - github-read.tool.ts
  - github-create-issue.tool.ts
  - github-sync.tool.ts
  - reconciliation-query.tool.ts

Each file exports a pure factory `createXxxTool(deps)` that takes a
typed `PlanningToolDeps` bag and returns a `defineTool({...})` result.
Tool names, parameter schemas, prompt snippets, and execute bodies are
preserved verbatim — only the `this.X` references are rewritten as
`deps.X`. The LLM's memorised tool surface is unchanged.

`src/modules/planning/tools/types.ts` defines `PlanningToolDeps` as a
typed interface listing every collaborator + callback the tool files
are allowed to touch. Adding a new dep means updating this interface
AND `PlanningService.buildToolDeps()`.

`src/modules/planning/tools/tool-registry.ts` exports
`createPlanningTools(deps)` which returns the array of 10 tools in
the original registration order.

`PlanningService` shrinks from 1016 to 577 lines (-439). Every
`createXxxTool()` method is gone. The `customTools: [...]` array in
`createOrResumeSession` is replaced by a single
`createPlanningTools(this.buildToolDeps())` call. The new
`buildToolDeps()` private method composes the dep bag with bound
callbacks for `emitStudioEvent` / `buildProposalDefaults` /
`getCurrentTurnId` so the SSE emitter + counter stay private to
PlanningService.

`planning.service.test.ts` rewires the github_create_issue tests
to construct the tool via `createGitHubCreateIssueTool(deps)`
directly (instead of the old `(service as any).createGitHubCreateIssueTool()`
prototype access). The dep bag is built from the same mocks the
test was already passing to PlanningService, so the existing
proposal-staging assertions exercise the real tool body.

Verified:
  - npm run check (tsc) — clean
  - npx vitest run --no-file-parallelism — 561/561 pass
  - npm run build:web — clean
  - PlanningService line count: 577 (target was < 600)
  - forwardRef count: 4 (unchanged)
  - grep for createXxxTool methods on PlanningService — 0 matches
  - PORT=3030 npm start boots, GET /health returns {"ok":true,...}

Phase 7 complete. Issue #227 ready for PR.
